requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
rank-bm25
rapidfuzz
numpy
python-dateutil